password-store extension for managing bitwarden cli session
Adds `pass bw` command which logs in or unlocks the bitwarden session as
necessary and stores the session in the encrypted password store. All
parameters are passed to `bw` after the session is validated or created.

Requires `PASSWORD_STORE_ENABLE_EXTENSIONS=true` to be set in the
environment to enable the per-user extensions.