tmux: explicitly set default-shell and default-command
Using tmux from HEAD to work around neovim sixel crash, and it now uses
the default shell for `run-shell`. With nu as the user's shell,
vim-tmux-navigator broke b/c it was using the nu `ps` builtin instead of
the os's ps. So explicitly set the default-shell to zsh or bash.
default-command is used for new panes - set that to $SHELL.